Add function for updating delta in selfish mining

+++ src/applications/model/selfish-miner-status.h
@@ -8,7 +8,7 @@ namespace blockchain_attacks{
         long MinedBlock;
         long SelfishMinerWinBlock;
         long HonestMinerWinBlock;
-        long Delta;
+        int Delta;
     } SelfishMinerStatus;
     
 }
